Update XY Plotter Auton.c
Previous code had speed control errors in the motors themselves which caused flicks at the end of diagonal lines shown in fig1. To fix this I added another speed calculation variable called "adjustedReducedSpeed" which adds a small speed increase proportional to the angle of the line. This adjustment had a huge improvement on the test lines drawn as seen in fig2. These lines no longer have the flick before the next line was drawn.