Update XY Plotter Auton.c
Added a "moveToTopRight" function to send the robot to the top right corner of the whiteboard. This is done so that the text that has just been written is more easily visible. I also hard-coded coordinates on the whiteboard for the letters of my name, although after doing so I realised I'd accidentally been using centimetres instead of millimetres and that I'd taken all my measurements from the edge of the whiteboard instead of the edge of the drawing area. To fix these issues I firstly added a multiplier of 10, as seen on line 17. This is used in the calculations on lines 19 and 20 to convert centimetres into millimetres. On lines 19 and 20 I also subtracted 9 cm from the x axis and 11 cm from the y axis (the minimum x and y values I'd hard coded) so that the drawing started from point (0,0) on the drawing area. These changes are temporary and only included for the purpose of testing the drawing speed control of the motors. When I ran the program I got the results seen in fig3. I am very pleased with the lines drawn and I am ready to move onto the next stage of programming letters that can be drawn in any order based on a user input.