Updated auton for user text input
Added arrays for X and Y coordinates of "A" "B" "C" and "1" characters (See fig4 for character design). These arrays are stored in separate files and added to the main program through a #include. Also added a basic text selector on the Vex brain LCD to allow the user to select text to print (fig 5). This means the robot is able to print any 10 character combination "A" "B" and "C" on the whiteboard (See fig6 for example). As well as this I wrote code that goes through the selected text and prints it. Moved the hard coded "DrawShaun" to a separate file as it was only for testing purposes. I also found the adjusted reduced motor speed had negative effects on the letters drawn so I removed that code and found it drew letters much better.