Fixed text input UI
In the previous version the text selector had an issue where the current character wasn't reset when the cursor moved making editing text very inconvenient as users would have to scroll backwards to type an "A". To fix this I added an "if" statement so that whenever the cursor moves the current letter is either set to "A" on a blank space or the letter already occupying the space. This makes it much easier and intuitive to use the text selector interface.