Added multi-line drawing
Modified the user interface to support multi line editing. Also moved the character drawing from task main to a separate function. This means it can be re-used to draw each of the 3 lines without having to be retyped. With this current program the robot is now able to draw the entire alphabet on multiple lines, although I decided not to add numbers as it complicates the UI and doesn't serve much purpose for most users.