Modified user interface
In the previous multi line code I found it frustrating to edit text on multiple lines as changing line would require the user to manually scroll all the way back to the start of the line to start editing. To fix this I added code that automatically moves the cursor back to the start of the line whenever the line is changed, which made it much easier to edit text.