Minor back-end changes
Added code that counts the widths of each character on the screen and stops the user from inputting any characters that go out of the draw-able area. Moved the trim right code (loop that deletes blank spaces from end of line) to it's own separate function so that it can be re-used for all 3 lines without having to be re-written. I found in previous code the move to right function actually blocked some of the printed text so to fix it I changed it to a move to bottom right function because it moves the drawing head to an unused area of the whiteboard and makes it easier to see the printed text.