Added option to repeat print
Added option to repeat print job after the robot has finishes printing by pressing button F-Up. This is done so that a user can input a string of text once then re-print it several times without having the re-enter it. There is also an option to end the print job by pressing button F-Down. Also found a bug in the previous code that when the user changed lines the current letter would remain on the previous letter they typed and not the letter the cursor is selecting. To fix this I modified user interface so that when the user changes lines the current letter is reset to whatever character they are selecting.